#30 Restructure event classes structure to match game logic better.
Created UI folder to hold all UI related events: screenEvents, toolEvents, gameControlEvents.
