+++ MultiplayerMod/Game/UI/GameSpeedControlEvents.cs
@@ -1,7 +1,7 @@
 using HarmonyLib;
 using MultiplayerMod.Core.Patch;
 
-namespace MultiplayerMod.Game.Events;
+namespace MultiplayerMod.Game.UI;
 
 [HarmonyPatch(typeof(SpeedControlScreen))]
 public static class GameSpeedControlEvents {
@@ -10,17 +10,20 @@ public static class GameSpeedControlEvents {
     public static event System.Action GameResumed;
     public static event SpeedChangedEventHandler SpeedChanged;
 
-    // ReSharper disable once InconsistentNaming
     [HarmonyPostfix]
     [HarmonyPatch(nameof(SpeedControlScreen.SetSpeed))]
+    // ReSharper disable once InconsistentNaming
+    // ReSharper disable once UnusedMember.Local
     private static void SetSpeedPostfix(int Speed) => PatchControl.RunIfEnabled(() => { SpeedChanged?.Invoke(Speed); });
 
     [HarmonyPostfix]
     [HarmonyPatch(nameof(SpeedControlScreen.OnPause))]
+    // ReSharper disable once UnusedMember.Local
     private static void OnPausePostfix() => PatchControl.RunIfEnabled(() => GamePaused?.Invoke());
 
     [HarmonyPostfix]
     [HarmonyPatch(nameof(SpeedControlScreen.OnPlay))]
+    // ReSharper disable once UnusedMember.Local
     private static void OnPlayPostfix() => PatchControl.RunIfEnabled(() => GameResumed?.Invoke());
 
 }
