+++ MultiplayerMod/Game/UI/Screens/ConsumableScreenEvents.cs
@@ -6,7 +6,7 @@ using HarmonyLib;
 using MultiplayerMod.Core.Patch;
 using UnityEngine;
 
-namespace MultiplayerMod.Game.Screens;
+namespace MultiplayerMod.Game.UI.Screens;
 
 public static class ConsumableEvents {
     public static event Action<string, string, bool> PermittedToMinion;
