+++ MultiplayerMod/Game/UI/Tools/Events/BuildEventArgs.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace MultiplayerMod.Game.Tools.Events;
+namespace MultiplayerMod.Game.UI.Tools.Events;
 
 [Serializable]
 public class BuildEventArgs : EventArgs {
