+++ MultiplayerMod/Game/UI/Tools/Events/CopySettingsEventArgs.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace MultiplayerMod.Game.Tools.Events;
+namespace MultiplayerMod.Game.UI.Tools.Events;
 
 [Serializable]
 public class CopySettingsEventArgs {
