+++ MultiplayerMod/Game/UI/Tools/Events/DragToolEvents.cs
@@ -5,7 +5,7 @@ using HarmonyLib;
 using MultiplayerMod.Core.Patch;
 using UnityEngine;
 
-namespace MultiplayerMod.Game.Tools.Events;
+namespace MultiplayerMod.Game.UI.Tools.Events;
 
 [HarmonyPatch]
 public static class DragToolEvents {
