+++ MultiplayerMod/Game/UI/Tools/Events/InterfaceToolEvents.cs
@@ -2,7 +2,7 @@
 using HarmonyLib;
 using UnityEngine;
 
-namespace MultiplayerMod.Game.Tools.Events;
+namespace MultiplayerMod.Game.UI.Tools.Events;
 
 [HarmonyPatch(typeof(InterfaceTool))]
 public static class InterfaceToolEvents {
