+++ MultiplayerMod/Game/UI/Tools/Events/UtilityBuildEventArgs.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace MultiplayerMod.Game.Tools.Events;
+namespace MultiplayerMod.Game.UI.Tools.Events;
 
 [Serializable]
 public class UtilityBuildEventArgs {
