+++ MultiplayerMod/Multiplayer/Commands/Tools/AbstractBuildUtilityCommand.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using MultiplayerMod.Game.Context;
-using MultiplayerMod.Game.Tools.Context;
-using MultiplayerMod.Game.Tools.Events;
+using MultiplayerMod.Game.UI.Tools.Context;
+using MultiplayerMod.Game.UI.Tools.Events;
 
 namespace MultiplayerMod.Multiplayer.Commands.Tools;
