+++ MultiplayerMod/Multiplayer/Commands/Tools/Modify.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using MultiplayerMod.Game.Context;
-using MultiplayerMod.Game.Tools.Events;
+using MultiplayerMod.Game.UI.Tools.Events;
 
 namespace MultiplayerMod.Multiplayer.Commands.Tools;
