+++ MultiplayerMod/Multiplayer/Configuration/GameEventBindings.cs
@@ -1,8 +1,8 @@
 ﻿using MultiplayerMod.Core.Dependency;
 using MultiplayerMod.Core.Logging;
-using MultiplayerMod.Game.Events;
-using MultiplayerMod.Game.Screens;
-using MultiplayerMod.Game.Tools.Events;
+using MultiplayerMod.Game.UI;
+using MultiplayerMod.Game.UI.Screens;
+using MultiplayerMod.Game.UI.Tools.Events;
 using MultiplayerMod.Multiplayer.Commands.Screens.Consumable;
 using MultiplayerMod.Multiplayer.Commands.Screens.Priorities;
 using MultiplayerMod.Multiplayer.Commands.Screens.Research;
@@ -65,15 +65,15 @@ public class GameEventBindings {
         ConsumableEvents.PermittedToMinion += (properName, consumableId, isAllowed) =>
             client.Send(new PermitConsumableToMinion(properName, consumableId, isAllowed));
 
-        ScheduleEvents.SchedulesChanged += schedules => client.Send(new ChangeSchedulesList(schedules));
+        ScheduleScreenEvents.SchedulesChanged += schedules => client.Send(new ChangeSchedulesList(schedules));
 
-        PrioritiesEvents.PersonalPrioritySet += (properName, choreGroup, value) =>
+        PrioritiesScreenEvents.PersonalPrioritySet += (properName, choreGroup, value) =>
             client.Send(new SetPersonalPriority(properName, choreGroup, value));
-        PrioritiesEvents.PersonalPrioritiesAdvancedSet += (value) =>
+        PrioritiesScreenEvents.PersonalPrioritiesAdvancedSet += (value) =>
             client.Send(new SetPersonalPrioritiesAdvanced(value));
 
-        SkillEvents.HatSet += (properName, hat) => client.Send(new SetHat(properName, hat));
-        SkillEvents.SkillMastered += (properName, skillId) => client.Send(new MasterSkill(properName, skillId));
+        SkillScreenEvents.HatSet += (properName, hat) => client.Send(new SetHat(properName, hat));
+        SkillScreenEvents.SkillMastered += (properName, skillId) => client.Send(new MasterSkill(properName, skillId));
     }
 
     private void BindUserMenu() {
